Updated LobbySearch to update the list of lobbies via subscription.
List of lobbies will grow or shrink whenever a lobby is made or removed.
The data relevant to the lobby is also updated through subscription (current player count and state).
It now properly sets lobbies to inactive when its state is 'IN_GAME' or 'PRIVATE'.
Alert message for trying to join an inactive game is updated.